don't filter out tracing logs during tests
