+++ .github/workflows/check.yml
@@ -33,7 +33,7 @@ jobs:
       uses: actions-rs/cargo@v1
       with:
         command: test
-        args: --verbose --all
+        args: --verbose --all -- --nocapture
     # Run clippy
     - name: Run clippy
       uses: actions-rs/cargo@v1
